For key-based paths, we're going to need to split the path based on dots. I'm not yet sure how I'm going to deal with square brackets for arrays, but let's get something down and some tests passing.
Once I've split the path up, my plan is to use a recursive method which steps through the path expression and data together. If we get to the end of the path, then we know we're done.

The test is just a simple case of one object nested inside another.