Now, go and fiddle on https://www.regexplanet.com/advanced/java/index.html until I have a good expression that will pick out either a key in the JSON path or an array reference each time I do find(). Using the test string 'a.a2[1].b[0][0]', I come up with the regex of (\[\d+])|([\w]+) i.e. either some digits in square brackets or some continuous word characters.
Then, I'm going to use Apex regex classes to call find() repeatedly on the given path expression each time I do a recursive call in my class. Which is basically acting like an iterator but Matcher.find() is not an iterator. Notably, find() both tests for more matches, and advances through the string. Whereas we'd expect an iterator's hasNext() not to mutate the state.

So, I take a side-track to wrap this Matcher functionality into a new class, RegexFindIterator. I'll make it a private inner class until I decide what to do with it.