RegexFindIterator seems like it can stand on its own, so make it into a top-level class.
Start to deal with some other considerations...

Going to need a code-path to deal with array indices, but leave it blank for now.

Also going to need to deal with missing keys. We should be able to return something other than null if a key is absent, so put something in for that and we'll add a way for users of the class to set it later.