First attempt at list handing...
If the current key has square braces in it, trim the first and last characters, and parse the rest as an integer. Then get either the corresponding item from the list, or missingKeyResult if it's out of bounds.

In doing this, I realise that I can pull nextData out and set it in either branch of testing what sort of key we have